[tuner] Refactor set_dispatch_tuner to accept codegen_pipeline instead of conv_lowering_strategy
